Move feature block text into pages
