+++ src/components/home/FeatureBlock.js
@@ -2,12 +2,18 @@ import React from 'react';
 import { TwoColumnLayout } from 'components';
 import { createUseStyles } from 'react-jss';
 
+import BackgroundImage from './BackgroundImage';
+
 const useStyles = createUseStyles((theme) => ({
   root: {},
 
   twoColumnLayoutRoot: {
     display: 'flex',
     flexDirection: 'column-reverse',
+  },
+
+  backgroundImage: {
+    height: 250,
   },
 
   [`@media (min-width: ${theme.breakpoints.values.md}px)`]: {
@@ -22,11 +28,15 @@ const useStyles = createUseStyles((theme) => ({
   },
 }));
 
-const FeatureBlock = ({ text, image, imageSide }) => {
+const FeatureBlock = ({ text, backgroundImageUrl, imageSide }) => {
   const classes = useStyles();
   const className = {
     root: classes.twoColumnLayoutRoot,
   };
+
+  const image = (
+    <BackgroundImage backgroundImage={backgroundImageUrl} className={classes.backgroundImage} />
+  );
 
   if (imageSide === 'right') {
     return (
