--- src/components/home/features/index.js
@@ -1,3 +1,0 @@
-export { default as FeatureCustomPlugins } from './CustomPlugins';
-export { default as FeatureServiceCatalog } from './ServiceCatalog';
-export { default as FeatureDocsLikeCode } from './DocsLikeCode';
